+++ libft/maths/maths.h
@@ -6,7 +6,7 @@
 /*   By: afaucher <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2014/01/04 13:28:15 by afaucher          #+#    #+#             */
-/*   Updated: 2014/01/07 08:35:09 by afaucher         ###   ########.fr       */
+/*   Updated: 2014/01/07 15:22:05 by afaucher         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,11 +18,11 @@
 # define ARIGHT(angle) (angle < PI / 2 || angle > 3 * PI / 2)
 # define AUP(angle) (angle < PI)
 # define ADOWN(angle) (angle > PI)
-float	ft_pow(float nb, int p);
-float	ft_abs(float nb);
+double	ft_pow(double nb, int p);
+double	ft_abs(double nb);
 long	ft_fact(long nb);
-float	ft_getrad(float rad);
-float	ft_cos(float rad);
-float	ft_sin(float rad);
-float	ft_tan(float rad);
+double	ft_getrad(double rad);
+double	ft_cos(double rad);
+double	ft_sin(double rad);
+double	ft_tan(double rad);
 #endif /* MATHS_H */
